added database storage for user assignment and exams

+++ src/main/java/com/example/educationsystem/DescriptiveQuestion.java
@@ -1,7 +1,7 @@
 package com.example.educationsystem;
 
 public class DescriptiveQuestion extends Question{
-    public DescriptiveQuestion(int lessonId, int examId, int questionId, double score, String question) {
+    public DescriptiveQuestion(int lessonId, int examId, int questionId, float score, String question) {
         super(lessonId, examId, questionId, score, question);
     }
 }
